requirements.txt
numpy>=1.20.0
scipy>=1.7.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
pandas>=1.3.0
opencv-python>=4.5.0
tensorflow>=2.8.0
torch>=1.10.0
torchvision>=0.11.0
albumentations>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
nibabel>=3.2.0  # medical imaging formats
pydicom>=2.2.0  # DICOM file handling
h5py>=3.6.0
SimpleITK>=2.1.0
deepdish>=0.3.6
tqdm>=4.62.0
flask>=2.0.0  # web interface
pytest>=6.2.5  # testing
sphinx>=4.2.0  # documentation
joblib>=1.1.0  # for parallel processing and caching
psutil>=5.9.0  # for monitoring system resources
gdown>=4.4.0  # for downloading large models/datasets
pyyaml>=6.0  # for configuration files
jinja2>=3.0.0  # for report templates
markdown>=3.3.0  # for report generation